pyproject.toml
[project]
name = "ai-data-analyst"
version = "0.1.0"
description = "AI Data Analyst with natural language to SQL"
requires-python = "==3.12"
dependencies = [
    "langgraph>=0.0.55",
    "sqlalchemy>=2.0.23",
    "pandas>=2.1.3",
    "plotly>=5.17.0",
    "streamlit>=1.28.1",
    "psycopg>=3.1.0",
    "openai==0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = true

[project.scripts]
start = "streamlit_app:main"

[tool.setuptools]
py-modules = ["config", "database", "streamlit_app", "visualization", "agent", "llm_handler"]